image-score/requirements.txt
# ---- Core data stack (kept compatible to avoid ABI issues) ----
pandas>=2.1,<2.3
numpy>=1.24,<2.0
pillow>=10.0
requests>=2.31
tqdm>=4.66
openpyxl>=3.1

# ---- Vision & OCR ----
# Use headless OpenCV (no GUI popups, best for scripts/servers)
opencv-python-headless>=4.8
pytesseract>=0.3.10
# OPTIONAL: pyzbar needs the system libzbar installed, or it will fail at runtime.
# On macOS: `brew install zbar`
# If you don't need it, leave it commented out.
# pyzbar>=0.1.9

# ---- CLIP + Torch ----
torch
torchvision
open-clip-torch
ftfy>=6.1.1
regex>=2024.5.15



